# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ETK API",
        "contact": {
            "name": "fbeck",
            "email": "loordy@live.ru"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        },
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        }
    ],
    "paths": {
...
        "/contract/approve/": {
            "put": {
                "tags": [
                    "Contract"
                ],
                "operationId": "/contract/approve/",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ApproveCodeApprove"
                },
                "responses": {
                    "200": {
                        "description": "Contract updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/ApproveCode"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Contract updated successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to update this lc"
                    },
                    "404": {
                        "description": "Contract not found"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contract"
                ],
                "operationId": "/contract/approve/",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ApproveCodeRequest"
                },
                "responses": {
                    "200": {
                        "description": "Contract saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/ApproveCode"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Contract retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to create this lc"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
...
        "/helpers/search/kodp": {
            "get": {
                "tags": [
                    "helpers"
                ],
                "operationId": "/helpers/search/kodp",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "Язык",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Поиск",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kodp retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
        "/helpers/kodp": {
            "get": {
                "tags": [
                    "helpers"
                ],
                "operationId": "/helpers/kodp",
                "responses": {
                    "200": {
                        "description": "Returns kodp",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
...
        "/helpers/search/nskz": {
            "get": {
                "tags": [
                    "helpers"
                ],
                "operationId": "/helpers/search/nskz",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Поиск",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nskz retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
        "/helpers/nskz": {
            "get": {
                "tags": [
                    "helpers"
                ],
                "operationId": "/helpers/nskz",
                "responses": {
                    "200": {
                        "description": "Returns nskz",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApproveCode": {
                "title": "НСКЗ",
                "description": "НСКЗ",
                "properties": {
                    "id": {
                        "title": "Ключ",
                        "description": "Порядковый номер записи в таблице",
                        "format": "int32"
                    },
                    "phone_number": {
                        "title": "Номер мобильного телефона",
                        "description": "Номер мобильного телефона",
                        "format": "int64"
                    },
                    "id_lc": {
                        "title": "id Контракта",
                        "description": "id Контракта",
                        "format": "int64"
                    },
                    "user_id": {
                        "title": "id Юзера запросившего подтверждение контракта",
                        "description": "id Юзера запросившего подтверждение контракта",
                        "format": "int64"
                    },
                    "user_id_approved": {
                        "title": "id Юзера подтвердившего контракт",
                        "description": "id Юзера подтвердившего контракт",
                        "format": "int64"
                    },
                    "updated_at": {
                        "title": "Дата обновления записи",
                        "description": "Дата обновления записи",
                        "format": "date"
                    },
                    "created_at": {
                        "title": "Дата создания записи",
                        "description": "Дата создания записи",
                        "format": "date"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "phone_number": 998977806493,
                    "id_lc": 1534573,
                    "user_id": 634,
                    "user_id_approved": 634,
                    "created_at": 1547199446,
                    "updated_at": 1547199762
                }
            }
        },
        "requestBodies": {
...
            "ApproveCodeRequest": {
                "description": "ApproveCode object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Запрос подтверждение контракта",
                                "description": "Запрос подтверждение контракта",
                                "required": [
                                    "id_lc",
                                    "phone_number"
                                ],
                                "example": {
                                    "id_lc": 1534573,
                                    "code": 995240
                                }
                            }
                        }
                    }
                }
            },
            "ApproveCodeApprove": {
                "description": "ApproveCode object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Подтверждение контракта",
                                "description": "Подтверждение контракта",
                                "required": [
                                    "id_lc",
                                    "code"
                                ],
                                "example": {
                                    "id_lc": 1534573,
                                    "phone_number": 998977806493
                                }
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "etk_auth": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}